refactor(account): group account-management code into a qubx.core.account_manager package
Moves the flat modules into a cohesive subpackage (mirrors the layout from #299):
- account_manager.py        -> account_manager/manager.py   (AccountManager + SimulationAccountManager)
- account_manager_config.py -> account_manager/config.py
- account_state.py          -> account_manager/state.py     (AccountState)
- order_state_machine.py    -> account_manager/state_machine.py

account_manager/__init__.py re-exports the public surface, so existing
`from qubx.core.account_manager import AccountManager` callers are unchanged. Direct
importers of the old flat modules are repointed at the package root (or the relevant
submodule for module-specific tests). events.py and connector.py stay in core/ — they
are the typed event bus and the connector contract, not account-manager internals.

Pure structural move + import repointing; no behavior change. Full suite 1712 passed.
