+++ src/qubx/core/account_manager/__init__.py
@@ -1,0 +1,30 @@
+"""Central account-management subsystem.
+
+- ``manager``       — AccountManager (state owner + state machine + reconcile) and the
+                      SimulationAccountManager backtest variant.
+- ``state``         — AccountState, the pure-data container the manager mutates.
+- ``state_machine`` — pure order-lifecycle transition rules (the bug-magnet, isolated).
+- ``config``        — AccountManagerConfig + the tick cron helper.
+
+Public symbols are re-exported here so callers import ``from qubx.core.account_manager
+import AccountManager`` without reaching into submodules.
+"""
+
+from qubx.core.account_manager.config import AccountManagerConfig
+from qubx.core.account_manager.manager import AccountManager, SimulationAccountManager
+from qubx.core.account_manager.state import AccountState
+from qubx.core.account_manager.state_machine import (
+    LEGAL_TRANSITIONS,
+    can_transition,
+    validate_transition,
+)
+
+__all__ = [
+    "AccountManager",
+    "SimulationAccountManager",
+    "AccountManagerConfig",
+    "AccountState",
+    "LEGAL_TRANSITIONS",
+    "can_transition",
+    "validate_transition",
+]
